+++ CHANGELOG.md
@@ -1,8 +1,24 @@
 # Changelog
 
+## V0.5 20.11.2025 Added Countdown
+
+- added Countdown in the format hh:mm:ss
+- added new styles
+
+## V0.4 18.11.2025, Added Stopwatch
+
+- added Stopwatch
+- minor style correction for nixie (colons)
+- drawed second separator in general
+
 ## V0.30 16.11.2025, Added alarm setting
 
-## V0.22 15.11.2025, Added numbers and hash styles
+- alarm to be set with --alarm=hh:mm
+
+## V0.22 15.11.2025, even more styles
+
+- added numbers style
+- added hash style
 
 ## V0.21 15.11.2025, Corrected version and VT320 style